requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
asyncpg==0.30.0
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
magic-filter==1.0.12
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
outcome==1.3.0.post0
pdf2image==1.17.0
pillow==11.1.0
propcache==0.2.1
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PySocks==1.7.1
python-dotenv==1.0.1
rich==13.9.4
selenium==4.28.1
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.37
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.18.3
